+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## New Features
 
 * Add setting to customize postpone menu items (#130)
+* Set data folder location via an environment variable (TASKUNIFIER_DATA_FOLDER) (#196)
 * New logging and debug tool
 
 ## Bug Fixes